test(schedules): extract fake host setup
